split and upgrade config processing
