+++ internal/config/cli_only_options.go
@@ -1,0 +1,6 @@
+package config
+
+type CliOnlyOptions struct {
+	ConfigPath string
+	Verbosity  int
+}
