+++ internal/config/development.go
@@ -1,0 +1,13 @@
+package config
+
+import "github.com/spf13/viper"
+
+type development struct {
+	ProfileCPU bool `mapstructure:"profile-cpu"`
+	ProfileMem bool `yaml:"profile-mem" json:"profile-mem" mapstructure:"profile-mem"`
+}
+
+func (cfg development) loadDefaultValues(v *viper.Viper) {
+	v.SetDefault("dev.profile-cpu", false)
+	v.SetDefault("dev.profile-mem", false)
+}
